+++ src/MLStyle.jl
@@ -2,6 +2,7 @@ module MLStyle
 
 # export Feature , @case, @data, @def, @match, Fun, (⇒), Pattern, Case, Failed, failed, PatternDef, pattern_match, app_pattern_match, (..), enum_next
 export @match, Many, PushTo, Push, Seq, Do, @data, @use, use, @used
+export defPattern, defAppPattern, defGAppPattern, mkPattern, mkAppPattern, mkGAppPattern
 
 include("Extension.jl")
 using MLStyle.Extension
@@ -13,30 +14,15 @@ include("toolz.jl")
 include("render.jl")
 
 include("MatchCore.jl")
-
-include("Pervasive.jl")
 using MLStyle.MatchCore
-using MLStyle.Pervasive
+
+include("Infras.jl")
+using MLStyle.Infras
+
+include("Pervasives.jl")
+using MLStyle.Pervasives
 
 include("DataType.jl")
 using MLStyle.DataType
-# include("utils.jl")
-# include("Feature.jl")
-
-# using MLStyle.Err
-
-# include("Match.jl")
-# using MLStyle.Match
-
-# include("ADT.jl")
-# using MLStyle.ADT
-
-# include("Infras.jl")
-
-# include("MatchExt.jl")
-# using MLStyle.MatchExt
-
-# include("Data/Data.jl")
-# using MLStyle.Data
 
 end # module
