+++ src/render.jl
@@ -24,8 +24,25 @@ function render(expr::Expr, config :: Dict{Symbol, Any}, nested=true)
     runAstMapper $ mapAst(hd_f, tl_f, expr)
 end
 
+
+
 function format(args, template)
-    constlist = [Expr(:call, :(=>), QuoteNode(arg), arg) for arg in args.args if arg isa Symbol]
+    function dispatch(arg :: Symbol)
+        Expr(:call, :(=>), QuoteNode(arg), arg)
+    end
+
+    function dispatch(arg :: Pair)
+        Expr(:call, :(=>), QuoteNode(arg[1]), arg[2])
+    end
+
+    function dispatch(arg :: Expr)
+        @assert arg.head == :(=)
+        sym = arg.args[1]
+        @assert arg isa Symbol
+        value = arg.args[2]
+        Expr(:call, :(=>), QuoteNode(sym), value)
+    end
+    constlist = map(dispatch, args.args)
     constlist = Expr(:vect, constlist...)
     config = Expr(:call, Dict{Symbol, Any}, constlist)
     Expr(:call, render, template, config)
